Data frame with one row per haul to DATRASraw alike object

Data frame with one row per haul to DATRASraw alike object

Usage

get_datrasraw(x)

Arguments

x

Data.frame. Zero filled catch data.

Examples

dat <- data.frame(
          species = 1:2,
          Year = 2016:2020,
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))
 ds <- split(dat,dat$species)
 lapply(ds, get_datrasraw)
#> $`1`
#> [[1]]
#> data frame with 0 columns and 0 rows
#> 
#> [[2]]
#>   species Year      lon      lat       sx       sy       covA      covB
#> 1       1 2016 174.6867 161.8105 44.39101 166.7030   2.509967  9.232774
#> 3       1 2018 162.2473 187.0538 47.79153 161.6585 -12.344579 31.642001
#> 5       1 2020 170.6115 163.6545 58.10136 168.4943   4.659531 27.093547
#> 7       1 2017 175.5153 157.9576 65.39622 153.3326  12.085511  7.388083
#> 9       1 2019 165.2868 179.7251 58.27111 168.7669  -3.728194 32.145961
#>       EFFORT haul.id
#> 1 -0.7218071       1
#> 3 18.4945698       2
#> 5 18.2526955       3
#> 7 28.3246879       4
#> 9 29.4373442       5
#> 
#> [[3]]
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "DATRASraw"
#> 
#> $`2`
#> [[1]]
#> data frame with 0 columns and 0 rows
#> 
#> [[2]]
#>    species Year      lon      lat       sx       sy      covA     covB   EFFORT
#> 2        2 2017 180.9993 171.2870 64.12220 166.5441 -4.123216 18.38055 38.11568
#> 4        2 2019 156.4036 175.1305 51.33300 187.1361  7.140378 17.26859 33.25724
#> 6        2 2016 157.2230 175.9432 43.73577 172.3330 12.783818 15.73533 21.72755
#> 8        2 2018 188.7450 184.4574 46.54834 169.7750 -4.742499  9.75035 37.19464
#> 10       2 2020 166.8239 173.1955 80.00056 168.0508  8.908936 16.31502 14.41384
#>    haul.id
#> 2        1
#> 4        2
#> 6        3
#> 8        4
#> 10       5
#> 
#> [[3]]
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "DATRASraw"
#>