Prediction a grid for a specific year

Prediction a grid for a specific year

Usage

get_prediction_grid(year, x, subsel = NULL, varsbyyr, vars)

Arguments

year

Numeric string. The years the data is available for.

x

Data.frame. The data.frame of the covariate data going into the model.

subsel

Default = NULL.

varsbyyr

Character string. The name of the variables that vary by year.

vars

Character string. The name of the variables that do not vary by year.

Value

list of data by year

Examples

dat <- data.frame(
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB2019 = rnorm(n = 10, mean = 20, sd = 10),
          covB2020 = rnorm(n = 10, mean = 20, sd = 10),
          covB2021 = rnorm(n = 10, mean = 20, sd = 10),
          covB2022 = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))

vars <- "covA"
varsbyyr <- "covB"
YEARS <- 2019:2022
pg <- lapply(YEARS,
       FUN = get_prediction_grid,
       x = dat,
       vars = vars,
       varsbyyr = varsbyyr)
names(pg) <- YEARS
pg
#> $`2019`
#>         lon      lat       sx       sy       covA      covB EFFORT
#> 1  177.9585 160.6156 45.84597 175.6965   1.544489 22.047814      1
#> 2  170.8536 153.9479 65.37068 170.6910  21.977080 17.264612      1
#> 3  160.9660 176.4863 55.12320 171.3167  17.013984 -0.900256      1
#> 4  169.0037 173.6666 62.15848 181.0926  -2.663429 15.410590      1
#> 5  169.4376 163.2860 58.28618 160.4927  13.917904 27.283537      1
#> 6  175.5241 156.5687 57.65729 191.8247   9.259522 28.899248      1
#> 7  173.6401 150.3188 68.47033 162.3948  13.555400 31.289390      1
#> 8  182.2847 168.4498 49.17068 178.3102  13.119209 17.993637      1
#> 9  167.0057 168.5081 63.64723 187.2335 -19.595259 18.215235      1
#> 10 152.5744 150.1972 44.23224 179.9299  -2.350154 32.705422      1
#> 
#> $`2020`
#>         lon      lat       sx       sy       covA      covB EFFORT
#> 1  177.9585 160.6156 45.84597 175.6965   1.544489 31.725575      1
#> 2  170.8536 153.9479 65.37068 170.6910  21.977080 26.299142      1
#> 3  160.9660 176.4863 55.12320 171.3167  17.013984 24.195563      1
#> 4  169.0037 173.6666 62.15848 181.0926  -2.663429  9.891661      1
#> 5  169.4376 163.2860 58.28618 160.4927  13.917904 38.700437      1
#> 6  175.5241 156.5687 57.65729 191.8247   9.259522 14.547017      1
#> 7  173.6401 150.3188 68.47033 162.3948  13.555400 30.456635      1
#> 8  182.2847 168.4498 49.17068 178.3102  13.119209 18.460641      1
#> 9  167.0057 168.5081 63.64723 187.2335 -19.595259 31.871781      1
#> 10 152.5744 150.1972 44.23224 179.9299  -2.350154 27.236926      1
#> 
#> $`2021`
#>         lon      lat       sx       sy       covA      covB EFFORT
#> 1  177.9585 160.6156 45.84597 175.6965   1.544489 15.746913      1
#> 2  170.8536 153.9479 65.37068 170.6910  21.977080 19.214249      1
#> 3  160.9660 176.4863 55.12320 171.3167  17.013984 19.435333      1
#> 4  169.0037 173.6666 62.15848 181.0926  -2.663429 24.867964      1
#> 5  169.4376 163.2860 58.28618 160.4927  13.917904  5.407470      1
#> 6  175.5241 156.5687 57.65729 191.8247   9.259522 12.727337      1
#> 7  173.6401 150.3188 68.47033 162.3948  13.555400 24.017640      1
#> 8  182.2847 168.4498 49.17068 178.3102  13.119209  8.249857      1
#> 9  167.0057 168.5081 63.64723 187.2335 -19.595259 31.212228      1
#> 10 152.5744 150.1972 44.23224 179.9299  -2.350154 17.122552      1
#> 
#> $`2022`
#>         lon      lat       sx       sy       covA      covB EFFORT
#> 1  177.9585 160.6156 45.84597 175.6965   1.544489 22.028656      1
#> 2  170.8536 153.9479 65.37068 170.6910  21.977080 11.529979      1
#> 3  160.9660 176.4863 55.12320 171.3167  17.013984 35.176391      1
#> 4  169.0037 173.6666 62.15848 181.0926  -2.663429 13.982099      1
#> 5  169.4376 163.2860 58.28618 160.4927  13.917904  8.458477      1
#> 6  175.5241 156.5687 57.65729 191.8247   9.259522 31.513010      1
#> 7  173.6401 150.3188 68.47033 162.3948  13.555400 26.081899      1
#> 8  182.2847 168.4498 49.17068 178.3102  13.119209 25.436888      1
#> 9  167.0057 168.5081 63.64723 187.2335 -19.595259 11.429476      1
#> 10 152.5744 150.1972 44.23224 179.9299  -2.350154 26.488005      1
#>