Data frame with one row per haul to DATRASraw alike object

Data frame with one row per haul to DATRASraw alike object

Usage

get_datrasraw(x)

Arguments

x

Data.frame. Zero filled catch data.

Examples

dat <- data.frame(
          species = 1:2,
          Year = 2016:2020,
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))
 ds <- split(dat,dat$species)
 lapply(ds, get_datrasraw)
#> $`1`
#> [[1]]
#> data frame with 0 columns and 0 rows
#> 
#> [[2]]
#>   species Year      lon      lat       sx       sy      covA     covB    EFFORT
#> 1       1 2016 150.4871 163.9425 66.49797 164.8145  7.059592 13.36828 35.405136
#> 3       1 2018 177.8019 173.2978 53.78261 166.7043  3.173622 15.27988 20.387656
#> 5       1 2020 167.3484 170.8283 46.33304 169.2488  9.818074 14.89254 17.752855
#> 7       1 2017 162.7013 180.6734 72.12597 168.8007 17.294662 33.95182  5.179582
#> 9       1 2019 185.2534 163.0928 43.99904 165.3589  7.106785 14.79454  4.924414
#>   haul.id
#> 1       1
#> 3       2
#> 5       3
#> 7       4
#> 9       5
#> 
#> [[3]]
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "DATRASraw"
#> 
#> $`2`
#> [[1]]
#> data frame with 0 columns and 0 rows
#> 
#> [[2]]
#>    species Year      lon      lat       sx       sy      covA     covB   EFFORT
#> 2        2 2017 165.6773 153.9855 57.69382 162.0112 14.588974 33.79577 31.02774
#> 4        2 2019 165.8345 181.9584 57.38854 175.7144 12.252396 11.99126 21.61437
#> 6        2 2016 171.9300 158.0902 63.34329 178.3322 11.345357 41.52126 16.32446
#> 8        2 2018 160.9925 169.6685 53.28805 152.5681  4.604156 29.41764 21.04013
#> 10       2 2020 183.2697 167.1471 64.13807 177.0909 -1.081134 29.05158 15.50017
#>    haul.id
#> 2        1
#> 4        2
#> 6        3
#> 8        4
#> 10       5
#> 
#> [[3]]
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "DATRASraw"
#>