Data frame with one row per haul to DATRASraw alike object

Data frame with one row per haul to DATRASraw alike object

Usage

get_datrasraw(x)

Arguments

x

Data.frame. Zero filled catch data.

Examples

dat <- data.frame(
          species = 1:2,
          Year = 2016:2020,
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))
 ds <- split(dat,dat$species)
 lapply(ds, get_datrasraw)
#> $`1`
#> [[1]]
#> data frame with 0 columns and 0 rows
#> 
#> [[2]]
#>   species Year      lon      lat       sx       sy        covA      covB
#> 1       1 2016 171.9658 185.9311 57.12508 153.4558  -0.6086003 28.108688
#> 3       1 2018 164.5690 161.0045 45.40534 168.9528  12.6279807 33.533712
#> 5       1 2020 176.2555 167.9341 70.82258 163.8001 -10.1355760  3.793549
#> 7       1 2017 176.9951 170.1377 62.57869 164.4629   2.8398308 26.125135
#> 9       1 2019 175.4708 174.2312 64.75088 165.0390   4.0540046 21.531879
#>     EFFORT haul.id
#> 1 45.73173       1
#> 3 14.19250       2
#> 5 14.78188       3
#> 7 18.07005       4
#> 9 13.94318       5
#> 
#> [[3]]
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "DATRASraw"
#> 
#> $`2`
#> [[1]]
#> data frame with 0 columns and 0 rows
#> 
#> [[2]]
#>    species Year      lon      lat       sx       sy       covA     covB
#> 2        2 2017 189.1875 162.9365 46.45449 181.0881  16.473323 12.49986
#> 4        2 2019 166.1214 163.1017 32.08388 168.4485   3.861784 25.02678
#> 6        2 2016 178.6857 164.7567 55.20306 150.6914 -11.871169 17.74088
#> 8        2 2018 164.8139 156.7186 67.43077 149.3575  19.667027 26.13118
#> 10       2 2020 161.5189 158.4678 51.99045 173.6875  12.443499 23.82653
#>      EFFORT haul.id
#> 2  11.20447       1
#> 4  21.80970       2
#> 6  21.91519       3
#> 8  18.41592       4
#> 10 11.94535       5
#> 
#> [[3]]
#> data frame with 0 columns and 0 rows
#> 
#> attr(,"class")
#> [1] "DATRASraw"
#>