Prediction a grid for a specific year

Prediction a grid for a specific year

Usage

get_prediction_grid(year, x, subsel = NULL, varsbyyr, vars)

Arguments

year

Numeric string. The years the data is available for.

x

Data.frame. The data.frame of the covariate data going into the model.

subsel

Default = NULL.

varsbyyr

Character string. The name of the variables that vary by year.

vars

Character string. The name of the variables that do not vary by year.

Value

list of data by year

Examples

dat <- data.frame(
          lon = rnorm(n = 10, mean = 170, sd = 10),
          lat = rnorm(n = 10, mean = 170, sd = 10),
          sx = rnorm(n = 10, mean = 60, sd = 10),
          sy = rnorm(n = 10, mean = 170, sd = 10),
          covA = rnorm(n = 10, mean = 3, sd = 10),
          covB2019 = rnorm(n = 10, mean = 20, sd = 10),
          covB2020 = rnorm(n = 10, mean = 20, sd = 10),
          covB2021 = rnorm(n = 10, mean = 20, sd = 10),
          covB2022 = rnorm(n = 10, mean = 20, sd = 10),
          EFFORT = rnorm(n = 10, mean = 20, sd = 10))

vars <- "covA"
varsbyyr <- "covB"
YEARS <- 2019:2022
pg <- lapply(YEARS,
       FUN = get_prediction_grid,
       x = dat,
       vars = vars,
       varsbyyr = varsbyyr)
names(pg) <- YEARS
pg
#> $`2019`
#>         lon      lat       sx       sy        covA     covB EFFORT
#> 1  179.4242 184.3270 62.10293 164.0738   6.2686226 21.87722      1
#> 2  170.9362 178.8298 61.52399 177.1213 -19.6268257 16.67652      1
#> 3  174.6364 155.5844 37.80219 173.4947  -8.0634191 27.31421      1
#> 4  177.4054 163.4338 71.60276 176.2007  10.8044773 42.63101      1
#> 5  165.8797 178.9615 58.29921 179.7318   0.2806729 16.20041      1
#> 6  171.0307 179.5228 46.62181 179.1862   7.5315336 27.22897      1
#> 7  189.8857 143.8038 57.02544 170.6029 -10.4098203 17.41273      1
#> 8  185.5979 179.0027 49.17806 183.0194  -9.6861084 24.26300      1
#> 9  175.3998 172.6949 55.97855 166.1138  22.7668549 16.61735      1
#> 10 158.2016 169.2381 67.60331 175.8735  11.0930703 25.07773      1
#> 
#> $`2020`
#>         lon      lat       sx       sy        covA      covB EFFORT
#> 1  179.4242 184.3270 62.10293 164.0738   6.2686226 10.964659      1
#> 2  170.9362 178.8298 61.52399 177.1213 -19.6268257 16.851065      1
#> 3  174.6364 155.5844 37.80219 173.4947  -8.0634191 11.002124      1
#> 4  177.4054 163.4338 71.60276 176.2007  10.8044773  9.086924      1
#> 5  165.8797 178.9615 58.29921 179.7318   0.2806729 19.759115      1
#> 6  171.0307 179.5228 46.62181 179.1862   7.5315336 21.045404      1
#> 7  189.8857 143.8038 57.02544 170.6029 -10.4098203 23.766747      1
#> 8  185.5979 179.0027 49.17806 183.0194  -9.6861084 20.378613      1
#> 9  175.3998 172.6949 55.97855 166.1138  22.7668549  2.262500      1
#> 10 158.2016 169.2381 67.60331 175.8735  11.0930703 31.034716      1
#> 
#> $`2021`
#>         lon      lat       sx       sy        covA      covB EFFORT
#> 1  179.4242 184.3270 62.10293 164.0738   6.2686226 22.678481      1
#> 2  170.9362 178.8298 61.52399 177.1213 -19.6268257  7.154168      1
#> 3  174.6364 155.5844 37.80219 173.4947  -8.0634191 20.151753      1
#> 4  177.4054 163.4338 71.60276 176.2007  10.8044773 17.286984      1
#> 5  165.8797 178.9615 58.29921 179.7318   0.2806729 19.649726      1
#> 6  171.0307 179.5228 46.62181 179.1862   7.5315336 25.220538      1
#> 7  189.8857 143.8038 57.02544 170.6029 -10.4098203 29.448537      1
#> 8  185.5979 179.0027 49.17806 183.0194  -9.6861084 37.051956      1
#> 9  175.3998 172.6949 55.97855 166.1138  22.7668549 15.029900      1
#> 10 158.2016 169.2381 67.60331 175.8735  11.0930703 19.500755      1
#> 
#> $`2022`
#>         lon      lat       sx       sy        covA      covB EFFORT
#> 1  179.4242 184.3270 62.10293 164.0738   6.2686226 16.368636      1
#> 2  170.9362 178.8298 61.52399 177.1213 -19.6268257 17.890457      1
#> 3  174.6364 155.5844 37.80219 173.4947  -8.0634191 12.690109      1
#> 4  177.4054 163.4338 71.60276 176.2007  10.8044773 15.144942      1
#> 5  165.8797 178.9615 58.29921 179.7318   0.2806729 13.501317      1
#> 6  171.0307 179.5228 46.62181 179.1862   7.5315336 18.367647      1
#> 7  189.8857 143.8038 57.02544 170.6029 -10.4098203 18.081750      1
#> 8  185.5979 179.0027 49.17806 183.0194  -9.6861084 22.082121      1
#> 9  175.3998 172.6949 55.97855 166.1138  22.7668549  3.429071      1
#> 10 158.2016 169.2381 67.60331 175.8735  11.0930703 19.480452      1
#>